Reworked entire project: - included references to sequencing tables as opposed to sequencer objects. - included additional support for managing programs. - added serial interface support to select: program, period - added support to present names of programs.

+++ lib/ledSequencer/Storage.hpp
@@ -13,10 +13,10 @@ private:
 public:
    Storage();
    void init();
-   int8_t getProgram();
-   void setProgram(int8_t);
-   int8_t getPeriod();
-   void setPeriod(int8_t);
+   uint16_t getProgram();
+   void setProgram(uint16_t);
+   uint16_t getPeriod();
+   void setPeriod(uint16_t);
 };
 
 #endif // __STORAGE__
